Add disc_mismatch lint
This lint detects when union case labels reference constants from the
wrong enum type. For example, if a union has discriminator of type Color
but uses a case label from the Size enum, this will now be detected.

The implementation leverages the Numeric::Const references we preserve
during evaluation to determine which enum each case label belongs to.